Android: Fix an UnsatisfiedLinkError in JNIWrappers.saveState()
This is typically caused if a native method of a DSO is called before it
is loaded -- this could happen if the onPause is called twice before the
app has fully loaded all its libraries.

This should fix the problem, but I'm not 100% sure since I can't
reproduce this on any of my devices.

(cherry picked from commit 0d5d4de1698d45806fa5977ac71bd79f6f4e475b)