Force-place the dialog over the parent window
For some reason, sometimes, a dialog is placed in the top-level
corner of the desktop by Qt or by the Window Manager, despite
adjustPosition in QDialog doing the right thing.

This copies the QDialog adjustPosition code to KoDialog (since
it's obviously not public) and uses a singleshot time to move
the dialog window to the right place after 3 ms from the showEvent.

BUG:412393
(cherry picked from commit a286d2383f7fa8938bdc3d3d4fa04506ded335d0)