Only access the clipboard when we really need it
* remove two superfluous sig/slot connections in KisClipboard.
These QClipboard modes are not relevant for Krita, and were
fired together with dataChanged, resulting in clipboardChanged
being called three times.
* remove the preview() method which was only used by the
custom image widget
* set the default w/h for the imagefromclipboard widget to 1,
so we don't need to retrieve the QImage from the clipboard just
to get the actual dimensions
* simplify KisClipboard::clipboardDataChanged so it doesn't need
to retrieve the qimage from the clipboard. This means that hasClip()
can be true because there are images on the qclipboard, so that needs
to be tested.

BUG:443183
(cherry picked from commit bdf06f6511a8c29a3988b021cedca0b23a0a74b7)