Fix a crash in undo stask
The crash happened between the consurrent access to the stack from two
actors:

1) One stroke adding its command to the undo stack and calling
   KUndo2QStack::checkUndoLimit() to purge the oldest command
   of the stack to satisfy the undo stack limit.

2) The user starting the new stroke, which calls
   KUndo2QStack::purgeRedoState() from the GUI thread in a non-
   synchronized way.

Obviously, the crash happens only when the undo stack limit is
reached. Otherwise both the functions do nothing and no crash
happens.

The fix of the bug just moves the call to purgeRedoState() from
the point when a stroke is added into the queue to a point when
the stroke is actually started by the queue.

Huge thanks to @mako and @acc4 from Krita Artists for their help
with testing all my packages and tracking the crash down :)

BUG:456470