Hide everything on the welcome page except the central widget.
This currently works as something like this:

[Stage: Initialization i.e nothing is visible]

1. When Krita starts, we load the `State` from the config, we restore
the MainWindow to use this state.

[Stage: ShowEvent]

2. Once we get this event, we save the initialized state of the window
and then stop the auto-saving of mainwindow because we're about to hide
everything.

[Stage: Launch Canvas]

3. Once we launch the canvas, we restart the auto-saving of mainwindow
and load the state which we saved on step 2.

[Stage: Closing]

4. If the window is closed when we're on the welcome page. We don't save
the state. What makes sure that we didn't lose the state when we're
closing here is that canvas->welcome-page guarantees that we will save
the state of dockers etc.

5. If the window is closed on canvas page, then we do the saving of
mainwindow state as usual.

(cherry picked from commit 9b198a89086d0fdd5e6df457bf0466c6033d48de)
(cherry picked from commit 9a87f2d336db7b017822b3816887063486b4391c)