Fix slowness of dragging gradient sliders
The sluggishness was a result of a redundant computation which came from:

`sigSelectedStop -> KisStopGradientEditor::stopChanged ->
KisStopGradientEditor::positionChanged -> KoStopGradient::setStop ->
updatePreview`.

(The setStop on previous line (previous to sigSelectedStop) already
calls updatePreview.)

We block the signal of stopEditor while we initialize the gradient
editor's position slider after the stop has changed.