Add Snap To Pixel to MyPaint brush editor
Before this commit, Snap To Pixel option was absent from
MyPaint brush editor in Krita despite it being in the libmypaint
that Krita uses. This commit adds it in the correct place.

The range should in theory be (0,1), but there are nice, random-like
textured results one can get when using Snap To Pixel with non-pixelated
brushes, so range 0-10 is useful.

To get the usual behaviour (for pixel art brushes), one should use value 1.0.

CCBUG:453054