Add Pressure Gain Log to MyPaint brush editor
Before this commit, Pressure Gain Log option was absent from
MyPaint brush editor in Krita despite it being in the libmypaint
that Krita uses. This commit adds it in the correct place.

The range in libmypaint is -1.8 to 1.8, but values above 1.8 are still useful.
Hence the range in Krita is from -2 to 3.

Link:
https://github.com/mypaint/libmypaint/blame/master/brushsettings.json#L774

Default is 0.0.

CCBUG:453054