Add Offset By Speed Filter option to MyPaint engine
iBefore this commit, Offset By Speed Filter option was absent from
MyPaint brush editor in Krita despite it being in the libmypaint
that Krita uses. This commit adds it in the correct place.

Note: this is dependant on Offset by Speed option.

Range and name (the internal one is `by Speed Slowness`, not `by Speed
Filter`) was taken from libmypaint.

CCBUG:453054