Fix MacOS Gestures
Previously, the gesture handling on MacOS was a bit of a mess. If we had
a GestureType in our config, it would conflict with MacOSGestureType.
Resulting in gestures working only when they're loaded before or after
others.

Now, this is sorted. MacOSGestureType is handled just like GestureType
is handled everywhere but in the serializer and unserialzer.

In unserialize:

We disable GestureType gestures from being loaded, but instead load
MacOSGestureType ones (this we load as GestureType because nobody else
is aware of the destinction between this).

In serialize:

We save GestureType gestures as MacOSGestureType. Because that's how
they're shown in UI due to our hack in unserialize.

BUG:456446