Android: Fix NullPointerException caused due to an invalid content Uri.
This happened if we had persisted permissions to a Uri through
takePersistablePermission() and deleted the directory. Then on the
successive run, when we use getPersistedUriPermissions, Android still
sends us the previous Uri -- even though it is invalid. And using it
causes a crash.

Link to bug: https://krita-artists.org/t/the-latest-version-of-android-krita-cannot-export-image-sequence/46342