Fix update link on Welcome Page widget
Before this commit, the update link on the Welcome Page widget
would look clickable but it wasn't possible to click it.
That was because the widget wasn't enabled.
This commit enables it and makes the link clickable.

BUG:458034


(cherry picked from commit 148a5e9bb784d4d84ff5595afe1ab9e48f70883d)