Make MyPaint eraser use the proper brush settings
When enabling eraser mode, Krita will use the default MyPaint brush
settings instead of the ones the user has actually selected. This is
unlike MyPaint, GIMP and Drawpile do it, which use the selected brush to
erase.

This is due to a setting that enables this odd behavior, which is not
actually exposed in the user interface, meaning there's no way to turn
it off. This commit outright removes that setting and makes Krita behave
properly: use the brush that's selected and erase with it, don't switch
to something else behind the scenes.

BUG: 445206


(cherry picked from commit 93bf757c4e4e2bfd9e53eb6b31c8117cd98303cf)