Add middle handles to perspective transform
In addition to the handles on the four corners, there's now also handles
in the middle of the sides. Dragging one of them is like moving both
corner handles next to them simultaneously, just like e.g. Paint Tool
SAI and Drawpile do it.

Renames some members from "corner" to "handle" point and adds an enum
for the handle indexes so that there's less magic numbers involved.