Fix the "Start Numbering At" Field not Working as Advertised
Internally, the code always refers to this as 'numbering offset' so I
presume that the labeling changed in the GUI and that wasn't properly
reflected in the code to work as intended. In short, I've modified the
behavior to work as advertised.

I plan to make some modifications to export at some point, but one of the
best things we could do short term is clean up variable naming to better
reflect intent as that is how this bug happened in the first place (I
think...)

BUG:458997