Reorganize access to display color converters
Putting a reference to the KoColorDisplayRendererInterface instance in
KisVisualColorModel was not the best decision, it just made splitting it
off of KisVisualColorSelector easier.
Now it's back in the selector widget, because the model really should
be independent of any UI.
The exposure control is now an optional duty of the selector, not the model.

Also, the docker now has a shared WGSelectorDisplayConfig that acts as
proxy for the current KisDisplayColorConverter and notifies all
widgets of changes.
This should be less messy than telling each affected widget individually
on each canvas switch.