Qt Android: Fix ANR caused by a deadlock
The deadlock happened between Android Main thread and qtMainLoopThread.

When a user would swift through context menu (QMenu) quickly, the
checkAndSetTopLevelWindow would get called several times on java thread
-- here the call was made to QPlatformWindow::hide() which resulted in
flushWindowSystemEvents getting called, and here we post a flushEvent
for qtMainLoopThread and then wait on a wait condition.

While all this is happening qtMainLoopThread is waiting for Android
thread to create a new surface for the context menu surface. (but since
the android/java thread is busy waiting for flush events) -- we enter a
deadlock.

So, in this commit we post the events to application (main) thread and
we assume that because the user events too are posted to mainThread, all
of them will run sequentially.

This should also prevent some crashes due to the same.

Link to bug: https://krita-artists.org/t/bug-app-soft-locks-itself-when-trying-to-navigate-the-top-menu/47305/10