Allow adding more than 1 blank keyframe
Allows to add blank animation layers as the selection.

I talked briefly with emmet about this and it could be a good addition for some workflow cases. However it needs a bit of UI testing, so before fixing the issues I need to test a bit if the change is indeed feels ok with the workflow design we have for animations.

As it is now I think it needs polish, i.e. If the current layer is locked and the selection goes beyond more than one layer no blank layer is added. if current layer is no locked it will ad blank frames on non locked frames. Which is inconsistent.

BUG:459455
 

- [x] I confirmed this builds.
- [ x] I confirmed Krita ran and the relevant functions work.
- [ x] I tested the relevant unit tests and can confirm they are not broken. (If not possible, don't hesitate to ask for help!)
- [ x] I made sure my commits build individually and have good descriptions as per [KDE guidelines](https://community.kde.org/Policies/Commit_Policy).
- [x] I made sure my code conforms to the standards set in the HACKING file.
- [ x] I can confirm the code is licensed and attributed appropriately, and that unattributed code is mine, as per [KDE Licensing Policy](https://community.kde.org/Policies/Licensing_Policy).