3rdparty: ensure the correct macOS architecture is built
Some packages need an explicit value of CMAKE_OSX_ARCHITECTURES to be
sent to allow a correct build. This is a bit of a hacky way of
achieving this. Currently CMake is able to properly build fat binaries
if "x86_64;arm64" flag is sent, however, internally CMake represents
lists as semicolon separated strings which makes it complicated to pass
the entire arch value to ExternalProject without turning into a list
in the process.

CCBUG: 428995