make "Krita Default" the default input profile when settings are reset
When settings are reset such that there is no current input profile
selected, "Krita Default" should be the initially selected profile.

In this scenario, Krita would normally select the first item in the
list of profiles, and up until recently, this approach has worked
because "Krita Default" has always been the first item in the list,
but that changed with the addition of the input profile
"Clip Studio Paint Compatible".