QPA Android: Revamp input dispatching
This patch fixes several issues with input dispatching on Android.

Primarily, we fix the locking up of the UI due to a deadlock between
Android UI Thread and Qt GUI thread. This also fixes an undefined state
between platform plugin and Qt's internal popup stack -- which resulted
in events not being received by the appropriate window.

Additionally, we change the handling of input events on to C++
side. This greatly simplifies the JNI interface and probably will make
things a lot easier if in future someone decides inputs should never
enter the java world.