Add several font-selection tests based on the web-platform-tests.
It also adds a test for Unicode Grapheme clusters, though I'd
really like to be able to create a test that checks for font-fallbacks.

I've got some notes for that written down now, I'll return to it later.

For the other tests, I managed to figure out how to get loading of an
svg file and then rendering it.
Opentype Variable fonts don't work right now, quirk of aggresive caching.