Fix nullptr dereference in pattern generator
If while dragging we click outside of the window bounds, none of the
pattern is selected resulting in:
widget()->patternChooser->currentResource() pointing to a nullptr. But
with includeHidden flag enabled we can still reference the last selected
pattern. So, this seems like a minor error (i.e operating on only "non
hidden" resource).

BUG:459906