Fix testParseFontStyles, remove "font" shorthand support.
The big problem with font shorthand, and all CSS shorthands
is that you sort of automagically need to figure out what
the different values mean. In particular here, font-family names
require special attention as they use quotes to indicate the spaces
are inside those quotes are part of the family name, and shouldn't
be considered a seperator. This will require some rethinking.