Fix baseline shift, spacing and tab spacing tests.
This also does some work for a new text decorations test,
but for some reason that works in Krita but not in the render test.
It also does work for the right-to-left test, but there's still
bugs there. See test-text-right-to-left-text-paths.svg, especially
in comparison to how webbrowsers render these. This bug will need
me to sit down and think very carefully, and is a little bit too
much for today.