Fix right-to-left test add right-to-left-anchoring test, fix anchoring.
This removes a hack that doesn't need to be there as long as we
default the first absolute text-transform to 0,0, which in turn
fixes some issues I had with bidi-reordered text.

Unfortunately I also found a peculiarity with regarding to bidi-
reordered text, where it can sometimes lead to holes in the different
chunks. Thinking it through, this seems to be what the spec intended,
of which in the other renderers, only firefox has the same behaviour.

I shouldn't forget to report this to the svgwg.