Android: Fix and workaround ForegroundServiceStartNotAllowedException
But since we get this exception, we catch it and try to workaround the
problem by starting the service as soon as the app is started and keep
it alive throughout (until the user closes the app).

When we enter onPause, this is when we send the saving command and try
to move the service in foreground, if we're successful -- well and good.
If not we continue using the service as is and try to finish (according
to the docs we have a couple of minutes -- but this also depends very
much on the memory pressure).

Once we're done with saving we use stopForeground() which removes the
notification and moves the service back to background, but we don't call
stopSelf here!

(cherry picked from commit ea2e00ddab98bbcb042cfd8171cbdfc93bf5ff18)