Bugfix: Make TIFF import and export work on Android
Prior to this we passed the filename to the libtiff and expect it to
open it. But this failed on Android, now we open the file using
QFile (since it supports content URI). And pass the file descriptor to
libtiff.

We again used filename for reading/writing exif data and since libexiv2
doesn't support using file descriptors we created a new IO backend which
is a wrapper around QFile. This was required since tiff file can be too
big to be put in the memory (had we used ioDevice->readAll()).

(cherry picked from commit 3c27e1d8a848320fc22f446ddde209ae25997f58)