Bugfix: Vector shapes not swapping the current fg/bg color.
It is really difficult to see why this used to work before. But to get
to the reason for the bug, let's say we press the swap button on paintop
box. And the current active widget is "Fill" tab. So, the
KoFillConfigWidget (d->fillVariant = Fill) is activated. So, when we
swap the color, the color of Stroke is changed through the color of
Fill, but the d->colorAction's color is that of the fill, but since we
are swapping it for Stroke. The if (currentKoColor() == color) returns
true and this we skip swapping it for stroke.

We also fix the colorAction button for Stroke, it didn't work
properly (for the same reason, Fill changed the color back to what it
was when swapping).

(A TODO for future (which maybe somewhere near, given some things are
still buggy with this file), we really have to decouple Stroke and Fill
types).

BUG:461692
(cherry picked from commit 2dfa0d19cd176ef70c082c4970fc76e1b0116900)