Add a plethora of missing autokey implementations for movement.
There are two problems right now (in general) with autokey compared to
previous versions of Krita.

1) Krita was originally heavily reliant on using KisTransactions for
autokeying content in a way that essentially inherited. Since dmitry
made some changes to the way transactions are applied during brush
strokes that allowed them to be multi threaded, this behavior had to be fixed
in a previous commit to prevent a data race resuling in a nasty crash on
autokey. This means we lost some of the "automatic" autokey behavior and
it needs to be reimplemented in a few key strategies in order to achieve
parity with older versions. In the case of this particular bug fix, I
missed a few edge cases when dealing with raster keyframes.

2) As a result of 1, there's a lot of areas of the code base where we're
definitely repeating ourselves. Long term, we should really try to come
up with a logical way to reuse any autokey related stroke strategy
behavior so that any stroke that needs to respect users autokey wishes
can do so with just a few function calls or configurations. For now,
there's no real issue with repeating ourselves just to fix regressions.

CCBUG:459723