Fix a multithreaded DB access in PredefinedBrushData's ctor
The original design of PredefinedBrushData was to be "always
intialized". But that is not very possible in some cases, e.g.
when we use this data object as a temporary representation of
the brush when loading it from XML in the non-GUI thread.

In such a case we cannot fetch the default resource from the
resources system, since we are in a non-GUI thread and have no
KisResourcesInterface, since the object should be default-
constructible.

To workaround this issue we just implement an 'effective-data'
pattern inside KisPredefinedBrushModel, which loads the fallback
resource in case there was none loaded before. This fix also fixes
a FIXME item in PredefinedBrushData's ctor, which ignored all the
properties coming from CommonData.