After much refactor, I finally managed to display the `KisResourceItemListView` horizontally on a single row with no scrollbars, scrollbuttons and NO vertical scrolling.
I had to move the display responsabilities for `KisResourceItemChooser` inside `KisResourceItemChooser` and `KisResourceItemListView` because a lot of the calling code would call `KisResourceItemChooser::itemView()` and override all styling done inside `KisResourceItemChooser` and `KisResourceItemChooser`.

I also made KisResourceItemListView automatically scale it's icons when in `IconStripHorizontal` not because of accesibility but because if the icons were taller than the container scrolling would be locked to vertical regardles of scroll properties (you can hide it, disable it but it would still scroll on mouse wheel).

There is some serious spaghetty code out there calling `KisResourceItemChooser`, if this sticks I will have to dive deeper.