Android: Make it possible to reset the default resource location
Android, doesn't provide a way to restore the default file location
using its native file picker. User *can* pick the folder in some cases
but it will be a content Uri not a unix file path. The only way a user
can restore to default is to type it out manually and we can't expect
them to know what it is, since it isn't something properly defined.