Split logic from KisCurveOptionData into a polymorphic class
That is needed to let MyPaint brush engine to have a different
set of sensors.

Basically, KisCurveOptionDataCommon now accepts an abstract
KisSensorPackInterface object that defines its behavior. The
sensors pack is stored in a lazy-shared structure, so, hopefully,
it will not be copied/allocated too often.

Now we have two non-intersectinc subsets of "curve option data"
hierarchies: 1) KisCurveOptionData; and 2) MyPaintCurveOptionData.
They both can be passed to a KisCurveOptionWidget2, but otherwise
they are not interchangable.