LayerBox: Properly handle the current and selected indexes.
Prior to the commit, if the user had only a couple of selected layers,
and they desected one, the focus (as in the current index) would not
change to the selected layer. This can be confusing for the users. So
if the user now desects the layer, focus is shifted to the previous
selected layer (this is a best guess, since internally Qt's selected
indexes use an append() method when a new index is selected).