LayerBox: Fix the bug with layer not getting selected
When a new layer was created we wouldn't select it due to
selectionModel()->hasSelection() clause -- which assumed that if there
was a already a layer selected then the new layer would clear selection
when it shouldn't.

In addition to this, there was a bug caused by currentChanged calling
setData(). This resulted in dataChanged() being called by a signal which
in turn called setCurrentIndex again but this time since no selection
flags were persisted this would update the selection when it shouldn't.

Now, we add a signal blocker since we're just adding a UserRole and the
only calls that seem to rely on this signal are updateUI -- which gets
called anyways. So, it should be safe, I assume.

BUG:462879