Revert "Disable the touch docker on Android"
This reverts commit 56456f59f898536a470bcf19832788aa8bb0d6bb.

Not having touch docker caused this bug. I don't exactly know the reason
why this happened in the first place and given the difficulty associated
with bugs involving events pipeline, I think it is best to leave it in
the same state as it was before. But maybe fix the touch docker to
disable or delegate the buggy actions.

BUG:463318
CCBUG:462848


(cherry picked from commit 8d96b64dcdd6a7e1dcd7783a7866e0d208a48772)