xsimd: Avoid using zip_* in xsimd >= 10
xsimd 10 introduces an undocumented, breaking change in the batch
zipping semantics. These are relied on in the RGBA interleavers,
when they replaced the old gather/scatter implementation.

To work around them, the old implementation was brought back,
surrounded with suitable guards, and the corresponding error pragma
was added to KoRgbaInterleavers.h.

BUG: 463219