Round MyPaint dab channels properly
Using KoColorSpaceMaths::scaleToA. Before this it got floored, which
caused a supposedly fully opaque brush to leave some alpha left over,
and even add a bit of alpha when drawing over fully opaque pixels.

Also updates the draw_dab.png used in the test because the different
rounding causes infinitesimal differences (that is, corrections) in the
alpha of the image.