KisDelayedSaveDialog: port to KoDialog and fix extra spacing
This commit simplifies and standardizes the layout of the delayed
load/save dialog. Porting it to the standard KoDialog class enables
mapping the custom buttons to known standard actions (Ok, Cancel, and
User1) while also being able to squish the dialog whenever the buttons
are hidden.

I've also fixed the dialog being still able to be closed via the X
button when the type is Forced.