Perl: fix deployment and environment setup
The path_list_append that was attempted for cmake -E env is a CMake 3.25
feature, while we only require 3.20 as a baseline.

Moreover, the usage of FindPerl meant that we had four separate
variables (PERL_EXECUTABLE, KRITA_PERL_EXECUTABLE,
KRITA_PERL_PATH_INTERNAL, and KRITA_PERL_PATH) for specifying something
that is already designated in the PATH
through PERL_BIN.

This commit removes all of them, uses a single point of truth (the PATH
in our build scripts), deploys Strawberry Perl (if required) to that
location, and checks that the PATH has been correctly carved out before
allowing the build to proceed.

CCMAIL: kimageshop@kde.org