GSL: readd Find Module to work around CMake pre 3.21.4
Prior to 3.21.4, CMake's Qt autogen support queried the imported
location of all link libraries. This is not possible when GSL is built
statically, because of a bug in the FindGSL module letting the static
MinGW library pass into the _DLL variables.

This was "fixed" in the following commit (the module bug remains):

https://github.com/Kitware/CMake/commit/895fa3433f50b7011fa475a4696dc06fe072aa25

To work around this, I'll readd the Find Module as written by Kitware
(to take advantage of the imported targets) but fix their mistaken
DLL logic.

CCMAIL: kimageshop@kde.org