KisAslLayerStyleSerializer: fix layer-local patterns
This is the counterpart of bug 441605, we only found this one out
because in 464700 Wolthera injects a placeholder pattern, which in turn
was never put into global storage.

For the sake of completionism, I've added several safe asserts in the
path that the nullptr makes, so that we get breakpointable warnings for
missing resources in the future.

BUG:464700
CCBUG:441605
CCMAIL:dimula73@gmail.com
CCMAIL:tamtamy.tymona@gmail.com