KisImportExportManager: ensure a color space with no profile is rejected on load
It was possible (unsure how, yet) to construct a color space that had no
profile in it. This would usually make it past the impex plugins (unless
there was a transform), but it would make the usage logger crash because
it records the profile name.

BUG:464848
(cherry picked from commit a5ad124f5d878f9bd249217432ffcaf6b93d12fb)