Resources: Add a centralized class for ThumbnailCaching
Across Krita there were several places which had thumbnail.scaled()
method, which was invoked over and over. The worst case of this was
ThumbnailPainter, where scaled() was called on all thumbnails for each
draw call making resizing of the docker quite slow.

The cache is a quite naive QMap of the QImages where the key is the
properties of the image and it implements no limits on how many image to
store and all that.