Fix issue with loading resources of the "basic tools" plugin
Now the basic tools pluging makes a static library that is then linked by the
main dynamic library of the plugin. The resource file was added to the static
library, but, according to Qt's documentation, using resources in a static
library may need the explicit use of Q_INIT_RESOURCE in the code
(https://doc.qt.io/qt-5/qdir.html#Q_INIT_RESOURCE). So this patch simply adds
the resources to the dynamic library.

Part-of: <https://invent.kde.org/graphics/krita/-/merge_requests/1730>