Make KisColorPatches use Model-View architecture for rendering color patches
There is one known bug that is present in the code which I couldn't figure out, when you first open Krita the Common colors widget responds to Kinetic scrolling in the direction it shouldn't be responding. For example if it is set to horizontal, then it will respond to scrolling in vertical direction. But if you change the direction or change the count of rows/columns, then things work like they should. I couldn't figure  out why this is happening,  something is changing headerLength when the widget is first opened but if you resize it, it snaps back to what it should be (which to me looks like a qt bug). 

BUG:463599