Change KisAssert to abort instead of exception
Dmitry's words:
"The original idea was to let unittests verify that
the code asserts on some specific test cases, but it
has never been used. So, I guess, it should now be safe
to just abort on unhandled assert."

This should fix nearly all the Coverity issues mentioning
throwing an exception in the destructor.

Part-of: <https://invent.kde.org/graphics/krita/-/merge_requests/1741>