Add assert in destructor in MultiChannelFilterBase
Before this commit, there was no assert in the destructor of
KisMultiChannelFilterBase because it was removed as a fix for
a Coverity issue regarding throwing exceptions in destructors.
This commit reverse that change, because it should be fixed in
KisAssert class, not by removing asserts. Also there is no need
to check the pointer for not being null before deleting.