Implement loading/saving of CSS Shapes for shape-inside/shape-subtract.
This loads url() style shapes (as used by inkscape), but also circle(),
ellipse(), path() and polygon(). All are saved as url().

For the test, it checks if the reference shape is the same. Additionally,
the shape-inside values are made visible in the text tool when hovering,
so we have a better idea what is happening there.