ui: Fix inconsistent misdetection of screen number
The reason why sometimes the widget change of screen wasn't detected was
because the check must be run on the main window, not on the canvas
widget.

The previous commit, as well as the use of screenNumber, lets me trace
all affected users of this check.

While at it, ensure that all known users of QApplication::activeWindow
in context with KisConfig::displayProfile won't crash if something goes
wrong.

BUG:466068