Fix disabled Link to Image in Ref Tool after drag-and-drop
Before this commit, there was no way to link the new reference image
to the original file, because Krita were forgetting the url.
This commit fixes that.

Notes:

- you can only choose Link to Image after adding the new reference
image. If you still use the Embed in .kra option and save and reopen,
then the location will be forgotten and there won't be a way to Link it
to Image again. (possible improvement: don't forget the location).

- in theory, the image displayed in Krita might be different before
and after saving and reopening, because before it will be loading the
file from clipboard, and the code there is pretty complicated (it
doesn't always go the same way as loading from url). To prevent issues,
just keep it as "Embed as .kra".

- removed "d->embed = false" in setFilename() and added
ref->setEmbed(false) in the loading visitor instead, because then the
responsibilities of parts of the code are more clear.

BUG:460058