Bugfix: Fix issue with paths that have similar content getting skipped
The bug:

Say, the user selected resource location was path = ~/.local/share/krita3 and
the AppDataLocation returns writableLocation = ~/.local/share/krita.
When we do path.startsWith(writableLocation) -- this returns true, when
it shouldn't because krita3 isn't built up on ~/.local/share/krita.

The solution:

Add a trailing separator to AppDataLocation, this will ensure that we
are comparing the right thing.