Don't forget 'align with layer' in gradient layer style
Before this commit, checking "Align with layer" in the gradient
overlay layer style would result only in temporary alignment
with the layer, and the checkbox would be unchecked next time
the user opens the dialog. The same was happening with the
gradient in the Stroke layer style.
This commit ensures the value of the checkbox will be remembered.

There is still potentially a problem with Stroke layer style
having only one boolean value for the "Align with layer" values in both
Gradient and Pattern, but it shouldn't be much trouble for the user
because this commit ensures that the value in the active stroke type
will be remembered correctly.

BUG:456194