Fix Pattern Overlay layer style "Link with layer" when moving
Before this commit, if you tried to move a layer with Pattern Overlay
with "Link with layer" checked, there would be an interesting effect of
pattern moving the exact opposite way of the layer.
This commit ensures it will move *with* the layer.

Note: it might potentially break existing files?