Refactor code to handle multiple lineboxes and line heights better.
Line height wasn't working properly before this, but it should
work now. With the new line and line chunk structs, we can now
generalize the line finalization code between flow-in-shape and
break-lines.

Bunch of inline size tests broke, but that's because they were
incorrect to begin with... Will fix them later as there's a few
more incorrect bits.