Implement basic wrapping area computation, add tests.
Krita now keeps track of shape-padding and shape-inside, and tries
to use them to compute the final wrapping shape. This is kinda slow
however, and we'll need to come up with a better solution.

Paragraph styles are now written separately, to avoid situations
where a tspan serializes non-tspan things like inline-size.

Similarly, this also updates the basic-svg-test to ensure we read
css units.

Some tests are added, but I'm hesitant to add the render tests just
yet.