Simplify content wrapper creation, various small fixes+tests.
This, for now, removes the content-wrapper code based on qpainterpath,
and simplifies it to QPolygonF, which is faster but cannot use padding
and margin. This will need a dedicated set of algorithms.

This also fixes a number of issues with text-indent, hanging-punctuation,
first-position selection, and adds test for indent, hanging-punctuation and
unexpected font-size changes.