Android: workaround multiple new files being created due to a race
This seems like an issue with event loop, so whenever the document
creation is successful, we accept the dialog (which should close it).
But before the dialog is closed, we enable its "Ok" button, this allows
makes it so the user can accidentally click it again and create multiple
files.

In my case, it was very easy to create two or three files, if the user
double taps, be it intentionally or accidentally.

The change moves the code so it is only activated if the image creation
somehow fails (which from the commit message, seems to be the original
purpose for enabling it).