Use device pixel ratio for selection outline width
Instead of messing around with physical screen size, since that breaks
when UI scaling gets involved. For example, on a 4K screen, you'd
probably run Krita with a scale factor of 2, which causes the entire UI
to be scaled, but the selection outline is impossible to make out
because it's just one pixel wide. With this change, it the outline just
scales like the rest of the user interface.

BUG:456364