Touch System: Cursor stuck as hand
This happened when the Touch Drag was mapped to Pan Canvas, and user
pressed the shortcut for Canvas Pan action. The reason for this is,
first the keyboard shortcut would activate the KisPanAction, and then
the Touch Shortcut would activate the action again, without deactivating
it first. The reason it is necessary is because setOverrideCursor in
KisPanAction maintains an internal stack for overridden cursor, so if we
call setOverrideCursor() twice, we have to restoreCursor() *twice*.

BUG:456183