Add a way to disable buildling translations at compilation step.
Building translation by themselves is not an issue, but since
translations add 20-30MiB to a package (APK), installing them takes
an extra 10 seconds (20 if wireless) to install.

The option is disabled by default and isn't enabled in any build script,
so should be up-to developer to toggle it or not.