Fix colorsmudge brush creating NaN values
This happened because xsimd:min, which calls vminq_f32 on ARM return NaN
if any of there parameter contains a NaN, which is different from
behavior of the instruction on x86.

Now, we create a mask out of nan values and for those indices set the
value to unitValue. This should work cross-platform.

BUG:467897