Fix height and width mixed up in color history
Before this commit, adjusting Height would adjust the width
of the color swatch, and adjusting the Width would adjust the height
of it.
Since it only changes the text on screen (other than that, the code
is consistent), it won't change the look of color selector
for anyone who already adjusted it to their liking.

BUG:444070